Extract mobile and data fetch logic into their own hook scripts
Trying to make `ImageGallery` have less logic and responsibilities to
hold.
